requirements.txt
aiogram
google-genai
redis